refactor registering ts-node
